pyproject.toml
[project]
name = "test-project"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.3",
    "sqlmodel>=0.0.38",
    "uvicorn[standard]>=0.44.0",
    "pydantic-settings>=2.0.0",
    "sqlalchemy>=2.0.0",
    "pymysql>=1.1.0",
    "cryptography>=41.0.0",
    "alembic>=1.13.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    "email-validator>=2.0.0",
    "python-multipart>=0.0.9",
    # 达梦数据库支持：
    # 1. sqlalchemy-dm 提供 SQLAlchemy dialect（pip install sqlalchemy-dm）
    # 2. dmPython 驱动需从达梦服务器安装包中获取，不在 PyPI 上
    #    安装方式：cd <dm_install_dir>/drivers/python/ && pip install dmPython
    # "sqlalchemy-dm>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.25.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
